+++ packages/client/src/links/core.ts
@@ -1,49 +1,24 @@
-import { AnyRouter, DataTransformer } from '@trpc/server';
-import { TRPCResult } from '@trpc/server/rpc';
+import { AnyRouter, DataTransformer, inferRouterError } from '@trpc/server';
+import { TRPCResponse } from '@trpc/server/rpc';
+import { observable } from '../rx/observable';
+import { Observable, Observer } from '../rx/types';
 import { TRPCClientError } from '../TRPCClientError';
 
-export type OperationContext = Record<string, unknown>;
+export type CancelFn = () => void;
+
+export type PromiseAndCancel<TValue> = {
+  promise: Promise<TValue>;
+  cancel: CancelFn;
+};
+
+export type OperationMeta = Record<string, unknown>;
 export type Operation<TInput = unknown> = {
   id: number;
   type: 'query' | 'mutation' | 'subscription';
   input: TInput;
   path: string;
-  context: OperationContext;
+  meta?: OperationMeta;
 };
-
-export type OperationResponse<TRouter extends AnyRouter, TOutput = unknown> =
-  | TRPCResult<TOutput>
-  | TRPCClientError<TRouter>;
-
-export type PrevCallback<TRouter extends AnyRouter, TOutput = unknown> = (
-  result: OperationResponse<TRouter, TOutput>,
-) => void;
-export type OperationLink<
-  TRouter extends AnyRouter,
-  TInput = unknown,
-  TOutput = unknown,
-> = (opts: {
-  op: Operation;
-  prev: PrevCallback<TRouter, TOutput>;
-  next: (
-    op: Operation<TInput>,
-    callback: PrevCallback<TRouter, TOutput>,
-  ) => void;
-  onDestroy: (callback: () => void) => void;
-}) => void;
-
-export type TRPCLink<TRouter extends AnyRouter> = (
-  opts: LinkRuntimeOptions,
-) => OperationLink<TRouter>;
-
-export interface HTTPLinkOptions {
-  url: string;
-}
-
-/**
- * @deprecated use `HTTPLinkOptions`
- */
-export type HttpLinkOptions = HTTPLinkOptions;
 
 export type HTTPHeaders = Record<string, string | string[] | undefined>;
 
@@ -57,15 +32,91 @@ export type TRPCFetch = (
 ) => Promise<Response>;
 
 export type LinkRuntimeOptions = Readonly<{
+  /**
+   * @deprecated to be moved to a link
+   */
   transformer: DataTransformer;
   headers: () => HTTPHeaders | Promise<HTTPHeaders>;
   fetch: TRPCFetch;
   AbortController?: typeof AbortController;
 }>;
 
-export type CancelFn = () => void;
+export interface OperationResult<TRouter extends AnyRouter, TOutput> {
+  data: TRPCResponse<TOutput, inferRouterError<TRouter>>;
+  meta?: OperationMeta;
+}
 
-export type PromiseAndCancel<TValue> = {
-  promise: Promise<TValue>;
-  cancel: CancelFn;
-};
+export type OperationResultObservable<
+  TRouter extends AnyRouter,
+  TOutput,
+> = Observable<OperationResult<TRouter, TOutput>, TRPCClientError<TRouter>>;
+
+export type OperationResultObserver<
+  TRouter extends AnyRouter,
+  TOutput,
+> = Observer<OperationResult<TRouter, TOutput>, TRPCClientError<TRouter>>;
+
+export type OperationLink<
+  TRouter extends AnyRouter,
+  TInput = unknown,
+  TOutput = unknown,
+> = (opts: {
+  op: Operation<TInput>;
+  next: (op: Operation<TInput>) => OperationResultObservable<TRouter, TOutput>;
+}) => OperationResultObservable<TRouter, TOutput>;
+
+export type TRPCLink<TRouter extends AnyRouter> = (
+  opts: LinkRuntimeOptions,
+) => OperationLink<TRouter>;
+
+export function createChain<
+  TRouter extends AnyRouter,
+  TInput = unknown,
+  TOutput = unknown,
+>(opts: {
+  links: OperationLink<TRouter, TInput, TOutput>[];
+  op: Operation<TInput>;
+}): OperationResultObservable<TRouter, TOutput> {
+  return observable((observer) => {
+    function execute(index = 0, op = opts.op) {
+      const next$ = opts.links[index];
+      const observer = next$({
+        op,
+        next(nextOp) {
+          const observer = execute(index + 1, nextOp);
+
+          return observer;
+        },
+      });
+      return observer;
+    }
+
+    const obs = execute();
+    const subscription$ = obs.subscribe(observer);
+
+    return () => {
+      subscription$.unsubscribe();
+    };
+  });
+}
+
+/** @internal */
+export function transformOperationResult<TRouter extends AnyRouter, TOutput>(
+  result: OperationResult<TRouter, TOutput>,
+) {
+  const { meta } = result;
+  if ('error' in result.data) {
+    const error = TRPCClientError.from<TRouter>(result.data);
+    return {
+      ok: false,
+      error,
+      meta,
+    } as const;
+  }
+  const data = (result.data.result as any).data as TOutput;
+  return {
+    ok: true,
+    data,
+    meta,
+  } as const;
+}
